requirements.txt
certifi==2020.12.5
chardet==3.0.4
idna==2.10
nose==1.3.7
Random-Word==1.0.4
requests==2.25.0
urllib3==1.26.2
